requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
dblpy==0.4.0
discord==1.0.1
discord.py==1.7.1
dnspython==2.1.0
humanfriendly==9.1
idna==2.10
multidict==5.1.0
pybase64==1.1.4
pymongo==3.11.3
python-dotenv==0.17.0
ratelimiter==1.2.0.post0
requests==2.25.1
selenium==3.141.0
six==1.15.0
splinter==0.14.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
